em-ai_api/pyproject.toml
[tool.poetry]
name = "em-ai-news-api"
version = "0.1.0"
description = ""
authors = ["ruo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "4.39.0"
datasets = "2.12.0"
bitsandbytes = "0.41.3"
accelerate = "0.25.0"
tokenizers = "0.15.2"
huggingface-hub = "0.20.2"
dill = "0.3.6"
multiprocess = "0.70.14"
pyarrow = "12.0.0"
responses = "0.18.0"
xxhash = "3.2.0"
pydantic = "1.10.10"
pandas = "^2.2.1"
elasticsearch = "^8.12.1"
elasticsearch-dsl = "^8.12.0"
langchain = {extras = ["all"], version = "^0.1.8"}
langchain-core = "^0.1.33"
langchain-elasticsearch = "^0.1.1"
langchain-openai = "^0.1.1"
lark = "^1.1.9"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
scipy = "^1.12.0"
peft = "0.5.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"
icecream = "^2.1.3"
openai = "^1.14.3"
flagembedding = "^1.2.8"
sqlalchemy = "^2.0.29"
requests = "^2.31.0"
pymysql = "^1.1.0"
aiohttp = "^3.9.3"
plantuml = "^0.3.0"

[tool.poetry.group.dev.dependencies]
black = "22.3.0"
flake8 = "4.0.1"
isort = "5.10.1"
py-spy = "^0.3.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
